test: Do not install Vbox Guest Additions
Currently, Virtualbox Guest Additions are installed to a test VM
image when it gets build by packer-ci. However, when we start a VM
on a Virtualbox which version does not match with the version
of the additions (i.e. != 5.2.0), either Vagrant or Vagrant's plugin
`vbguest` re-installs the additions which includes `mount.vboxsf`.

The latter is not compatible with the custom vboxsf version we use
in ubuntu-next and makes the following command to fail:

    mount -t vboxsf -o uid=1000,gid=1000 vagrant /vagrant

Removing the `mount.vboxsf` fixes the problem. However, I couldn't
find an easy way to remove the binary before Vagrant starts
syncing shared folders, thus I chose not to install the additions.

Signed-off-by: Martynas Pumputis <m@lambda.lt>