k8s: Retry CNP status update forever when Update() fails
The existing logic gave up after 5 retries and let the controller run another
cycle. This causes the exponential backoff to reset.

Instead, only return an error if the status update fails due to an error which
is less likely to be temporary. Such errors should be visible in the controller
status. However, errors due to Update() failing against the apiserver can be
retried forever using the exponential backoff. It is likely caused by resource
contention. If the caller wishes to impose a timeout, it can be done using the
context.

This also allows to provide metrics on the update duraton, success rate and
indication of average attempts required

Signed-off-by: Thomas Graf <thomas@cilium.io>