bpf: Use iptables TPROXY and shared proxy listeners
Change all Cilium proxies to use transparent redirect (TPROXY) to
shared proxy ports. This allows IP addresses and ports to remain
unchanged on redirected packets and removes the need to use a bpf map
to get the original destination address and port. Thus the proxymap is
completely removed, which helps reduce complexity of the bpf datapath.

Since the bpf socket helper is not yet available, we insert a set of
iptables rules for proxy redirects after a proxy has been configured.

Proxies still need to get the source security ID. This is done with a
lookup in the Cilium conntrack bpf maps. Lookup support is added for
both local and global CT maps.

Linux conntrack interferes with TPROXY as conntrack can NAT the traffic
due to 5-tuple collisions while TPROXY may use the non-NATted 5-tuple
for socket matching. This can lead to TCP 3-way handshake never finishing
due to TPROXY directing all packets to the listening socket, instead
of associating non-SYN packets with the new socket. This is resolved
by adding NOTRACK rules for proxy traffic.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>